pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.132.0",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
